flask-backend/requirements.txt
Flask
flask_cors
pymongo
python-dotenv
PyJWT
mysql-connector-python
flask
